Create AppDaemon_App.py
This is an example App Daemon app that runs on start & every 30 minutes to update 2 mqtt sensors in home assistant with the number of days in your current streak, and if your streak has been extended today or not (true/false). Requires manual yaml configuration of the 2 MQTT sensors in the HA config that listen on this topic.